* Move update_debug_logs back to container. How to access Editor.load?
squash
