+++ spyder/plugins/application/container.py
@@ -13,6 +13,7 @@ Holds references for base actions in the Application of Spyder.
 # Standard library imports
 import os
 import sys
+import glob
 
 # Third party imports
 from qtpy.QtCore import Qt, QThread, QTimer, Signal, Slot
@@ -26,7 +27,8 @@ from spyder.api.translations import get_translation
 from spyder.api.widgets.main_container import PluginMainContainer
 from spyder.config.utils import is_anaconda
 from spyder.plugins.console.api import ConsoleActions
-from spyder.utils.qthelpers import start_file, DialogManager
+from spyder.config.base import get_conf_path
+from spyder.utils.qthelpers import start_file, DialogManager, add_actions
 from spyder.widgets.about import AboutDialog
 from spyder.widgets.dependencies import DependenciesDialog
 from spyder.widgets.helperwidgets import MessageCheckBox
@@ -173,6 +175,20 @@ class ApplicationContainer(PluginMainContainer):
         # Debug logs
         self.menu_debug_logs = self.create_menu(
             LogsMenus.DebugLogsMenu, _("Debug logs"))
+        self.menu_debug_logs.aboutToShow.connect(self.update_debug_logs)
+
+    def update_debug_logs(self):
+        """Create an action for each lsp and debug log file."""
+        self.menu_debug_logs.clear()
+        debug_logs = []
+        files = glob.glob(os.path.join(get_conf_path('lsp_logs'), '*.log'))
+        files.append(os.environ['SPYDER_DEBUG_FILE'])
+        for f in files:
+            action = self.create_action(
+                f, f, triggered=lambda : None)  # TODO: add triggered
+            action.setData(f)
+            debug_logs.append(action)
+        add_actions(self.menu_debug_logs, debug_logs)
 
     def update_actions(self):
         pass
@@ -370,9 +386,9 @@ class ApplicationContainer(PluginMainContainer):
     @Slot()
     def restart_debug(self):
         answer = QMessageBox.warning(self, _("Warning"),
-             _("Spyder will restart in debug mode: <br><br>"
-               "Do you want to continue?"),
-             QMessageBox.Yes | QMessageBox.No)
+            _("Spyder will restart in debug mode: <br><br>"
+              "Do you want to continue?"),
+            QMessageBox.Yes | QMessageBox.No)
         if answer == QMessageBox.Yes:
             os.environ['SPYDER_DEBUG'] = '3'
             self.sig_restart_requested.emit()
